Remove on/off message
Minor nit, but whenever a buffer is loaded, the mode message that SVG
is on/off is displayed. While innocuous, I have a program that loads a
file and displays a message from a hydra, and the message is
overridden.

Most minor modes don't send such a message, this PR removes these
`message` calls and addresses Issue #29.