Revert "Added a match-data on tag that can be retrieved by callback"
This reverts commit 0e0ea48799d8911ed6c1ef60565a20fd5cf3dae4.

At the time `match-string' is evaluated it, at least sometimes,
returns nil (I suspect it always does), which then causes
`substring-no-properties' to error because it expects a string
as argument.

Alternatively, if this happens to actually return non-nil in
some cases, then `match-string-no-properties' should be used
instead, because, like `match-string', that won't error when
the requested match-data is empty.